TGL-8.feat: created types folder, added typesUser.d.ts
