+++ src/types/types.d.ts
@@ -1,0 +1,16 @@
+// Errors & validation messages
+export interface ValidationError {
+  errors: {
+    [key: string]: {
+      name: string;
+      message: string;
+    };
+  };
+  message: string;
+  name: string;
+  _message: string;
+}
+
+export interface GlobalError {
+  error: string;
+}
