+++ src/types/typesUser.d.ts
@@ -1,5 +1,3 @@
-// Models
-
 export interface User {
   _id: string;
   email: string;
@@ -42,21 +40,4 @@ export interface LoginMutation {
 // Api responses
 export interface RegisterResponse {
   user: User;
-}
-
-// Errors & validation messages
-export interface ValidationError {
-  errors: {
-    [key: string]: {
-      name: string;
-      message: string;
-    };
-  };
-  message: string;
-  name: string;
-  _message: string;
-}
-
-export interface GlobalError {
-  error: string;
 }
